requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-filter==22.1
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pip==22.2.2
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools==65.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
wheel==0.37.1
celery==5.1.2